Sync 425 commits of upstream lvgl/lvgl (Jan 12 -> Jun 15 2026) onto the
HarfBuzz feature branch ahead of resuming PR #9756.

Conflict resolutions:
- Kconfig / lv_conf_template.h: keep both LV_USE_HARFBUZZ and the new
  upstream LV_FREETYPE_CACHE_FT_GLYPH_L1 options.
- lv_freetype_private.h: give the HarfBuzz cache fields and upstream's
  glyph_l1 pointer separate #if/#endif blocks.
- lv_freetype.c (cache_node free): run both prerender/HarfBuzz teardown
  and upstream's lv_freetype_glyph_l1_deinit().
- lv_freetype_glyph.c: keep both GID_TAG macro and upstream's L1 cache
  macros. Verified the unicode L1 path (keyed by real codepoint) and the
  GID L2 path (GID_TAG-keyed) do not collide.
- src/lv_conf_internal.h: take upstream's deprecation shim; the generated
  config moved to include/lvgl/config/lv_conf_internal.h, regenerated via
  scripts/lv_conf_internal_gen.py so it carries LV_USE_HARFBUZZ.
- docs/.../font_support/index.rst: accept upstream deletion (docs migrated
  to .mdx; HarfBuzz doc to be ported separately).

Fix: lv_freetype_harfbuzz.h was the only file still including the
deprecated ../../lv_conf_internal.h; switched to ../../../include/lvgl/lvgl.h
to match sibling lib headers post-restructure.

Verified: syntax-clean (gcc -fsyntax-only) for FREETYPE+HARFBUZZ,
FREETYPE-only, and default (both off) configurations.

+++ src/draw/sw/lv_draw_sw_blur.c
@@ -13,10 +13,7 @@
 
 #if LV_USE_DRAW_SW
 
-#include "../../misc/lv_math.h"
-#include "../../misc/lv_types.h"
 #include "../../core/lv_refr_private.h"
-#include "../../stdlib/lv_string.h"
 
 /*********************
  *      DEFINES
@@ -235,7 +232,7 @@ void lv_draw_sw_blur(lv_draw_task_t * t, const lv_draw_blur_dsc_t * dsc, const l
             }
 
             uint8_t * buf_line_end = lv_draw_buf_goto_xy(t->target_layer->draw_buf, x_end, y);
-            blur_1_bytes_init(sum, buf_line_end, sample_len_limited, - px_size * skip_cnt);
+            blur_1_bytes_init(sum, buf_line_end, sample_len_limited, -(int32_t)px_size * skip_cnt);
             buf_prev = buf_line_end[0] + 1; /*Make sure that it's not equal in the first round*/
 
             for(x = x_start; x <= x_end; x += skip_cnt) {
@@ -324,7 +321,7 @@ void lv_draw_sw_blur(lv_draw_task_t * t, const lv_draw_blur_dsc_t * dsc, const l
             }
 
             volatile uint8_t * buf_line_end = lv_draw_buf_goto_xy(t->target_layer->draw_buf, x_end, y);
-            blur_3_bytes_init(sum, buf_line_end, sample_len_limited, - px_size * skip_cnt);
+            blur_3_bytes_init(sum, buf_line_end, sample_len_limited, -(int32_t)px_size * skip_cnt);
 
             for(x = x_start; x <= x_end; x += skip_cnt) {
                 blur_3_bytes(sum, buf_line_end, intensity);
